refactor: simplify pre-commit configuration for reliability
